+++ resources/lang/ca/settings.php
@@ -45,6 +45,10 @@ return [
     
     'reset_assistant' => "Assistent de reinici",
     'reset_assistant_help' => "Ací podràs escollir reiniciar les principals propietats dels estudiants",
-
+    
+    'themes' => "Temes i avatars",
+    'themes_disabled' => "Desactivar pacs d\'icones (utilitza <span class='tag is-dark'>Ctrl</span> per a seleccionar o deseleccionar)",
+    'custom_pack' => "Crea el teu pack d'imatges",
+    'custom_upload' => "Puja una imatge",
     
 ];